// src/app/models/Quotation.model.ts
export interface QuotationModel{
    IdTra_Quotation?: Number;
    Area_Quotation: string;
    Serie_Quotation: string;
    Date_Quotation: Date;
    Service_Id: Number;
    Company_Id: Number;
    Client_Id: Number;
    Status_Quotation: string;
    Document_Quotation: string;
    Update_Quotation?: Date;
    Validez_Quotation: Number;
    Total_Quotation: Number;
    SRestante_Quotation:Number;
}